### 2024-07-24 #7030 @danielskatz
👋 @goualard-f - Thanks for your submission. While we get started, you could work on the missing DOI that editorialbot suggests, which appears to me to be correct. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Additionally, I see a bunch of extra {}s in your bib file which are preserving cases, but these should be determined by the paper generator, not fixed (unless needed, such as for proper names). Can you remove the extra {}s?
