+++ README.md
@@ -42,5 +42,3 @@ Note that the `matplotlib` Python package must be available through `PyCall`.
 ## Project status
 
 The package is developed and has been tested on Julia 1.6 through 1.10.
-
-[![codecov.io](http://codecov.io/github/goualard-f/MicroFloatingPoints.jl/coverage.svg?branch=main)](http://codecov.io/github/goualard-f/MicroFloatingPoints.jl?branch=main)
