### 2024-07-24 df94bd4
Corrected joss submission according to preliminary suggestions.
