+++ JOSS2024/paper.md
@@ -22,7 +22,7 @@ The IEEE 754 standard defines the representation and the properties of the float
 
 # Statement of need
 
-Proving the properties of numerical algorithms involving floating-point numbers can be a very challenging task. Insight can often be gained by executing systematically the algorithm under study for all possible inputs. There are, however, too many values to consider with the classically available types `Float32` and `Float64`. Hence the need for libraries that offer smaller IEEE 754-compliant types to play with. SIPE [@lefevreSIPESmallInteger2012], FloatX [@flegarFloatXLibraryCustomized2019], and CPFloat [@fasiCPFloatLibrarySimulating2023], to name a few, are such libraries. However, being written in languages such as C or C++, they lack the interactivity and tight integration with graphical facilities that can be obtained from using script languages such as Julia. `MicroFloatingPoints.jl` is a Julia library that fills this gap by offering a parametric type `Floatmu` that can be instantiated to simulate in software small floating-point types: `Floatmu{8,23}` is a type using 8 bits to represent the exponent and 23 bits for the fractional part, which is equivalent to `Float32`; `Floatmu{8,7}` is equivalent to the Google Brain `bfloat16` format, ... `MicroFloatingPoints.jl` was, for example, instrumental in our understanding of the flaws of the algorithms computing random floating-point numbers by dividing a random integer by a constant [@goualardGeneratingRandomFloatingPoint2020]: being able to execute the algorithms for all possible inputs and to quickly display the results graphically in [Jupyter](https://jupyterbook.org/en/stable/intro.html) gave us the impetus to demonstrate rigorously that such procedures cannot ensure an even distribution of the bits in the fractional parts of the random floats, rendering them useless for applications such as *differential privacy* [@dworkDifferentialPrivacy2006; @mironovSignificanceLeastSignificant2012]. Figure \ref{fig:random} exemplifies the kind of result easily obtainable with `MicroFloatingPoints.jl`: using the type `Floatmu{7,16}`, we systematically divide *all* integers in $[0,2^{17}-1]$ by $2^{17}$ to obtain floating-point numbers in $[0,1)$. The picture shows the probability of being `1` for each bit of the fractional part of the result. 
[…]
