+++ JOSS2024/paper.bib
@@ -14,20 +14,23 @@
 	pages = {21--26},
 }
 
-@article{lefevreSIPESmallInteger2012,
-	title = {{SIPE}: {Small} {Integer} {Plus} {Exponent}},
-	shorttitle = {{SIPE}},
-	url = {https://hal.inria.fr/hal-00763954},
-	language = {en},
-	urldate = {2020-06-22},
-	author = {Lefèvre, Vincent},
-	month = dec,
-	year = {2012},
+@inproceedings{lefevreSIPESmallInteger2013,
+	address = {USA},
+	series = {{ARITH} '13},
+	title = {{SIPE}: Small Integer Plus Exponent},
+	isbn = {978-0-7695-4957-6},
+	doi = {10.1109/ARITH.2013.22},
+	booktitle = {Proceedings of the 2013 {IEEE} 21st Symposium on Computer Arithmetic},
+	publisher = {IEEE Computer Society},
+	author = {Lefevre, Vincent},
+	month = apr,
+	year = {2013},
+	pages = {99--106},
 }
 
 
 @article{rumpIEEE754PrecisionBase2016,
-	title = {{IEEE754} {Precision}-\$k\$ base-\${\textbackslash}beta\$ {Arithmetic} {Inherited} by {Precision}-\$m\$ {Base}-\${\textbackslash}beta\$ {Arithmetic} for \$k {\textless} m\$},
+	title = {{IEEE754} Precision-\$k\$ base-\${\textbackslash}beta\$ Arithmetic Inherited by Precision-\$m\$ Base-\${\textbackslash}beta\$ Arithmetic for \$k {\textless} m\$},
 	volume = {43},
 	issn = {0098-3500},
 	url = {http://doi.org/10.1145/2785965},
@@ -43,15 +46,12 @@
 
 
 @article{flegarFloatXLibraryCustomized2019,
-	title = {{FloatX}: {A} {C}++ {Library} for {Customized} {Floating}-{Point} {Arithmetic}},
+	title = {{FloatX}: A {C++} Library for Customized Floating-Point Arithmetic},
 	volume = {45},
 	issn = {0098-3500, 1557-7295},
-	shorttitle = {{FloatX}},
-	url = {https://dl.acm.org/doi/10.1145/3368086},
 	doi = {10.1145/3368086},
 	language = {en},
 	number = {4},
-	urldate = {2020-05-07},
 	journal = {ACM Transactions on Mathematical Software},
 	author = {Flegar, Goran and Scheidegger, Florian and Novaković, Vedran and Mariani, Giovani and Tom´s, Andrés E. and Malossi, A. Cristiano I. and Quintana-Ortí, Enrique S.},
 	month = dec,
@@ -61,7 +61,7 @@
 
 
 @article{highamSimulatingLowPrecision2019,
-	title = {Simulating {Low} {Precision} {Floating}-{Point} {Arithmetic}},
+	title = {Simulating Low Precision Floating-Point Arithmetic},
 	volume = {41},
 	issn = {1064-8275},
 	url = {https://epubs.siam.org/doi/abs/10.1137/19M1251308},
@@ -92,12 +92,12 @@
 
 @inproceedings{dworkDifferentialPrivacy2006,
 	address = {Berlin, Heidelberg},
-	series = {Lecture {Notes} in {Computer} {Science}},
-	title = {Differential {Privacy}},
+	series = {Lecture Notes in Computer Science},
+	title = {Differential Privacy},
 	isbn = {978-3-540-35908-1},
 	doi = {10.1007/11787006_1},
 	language = {en},
-	booktitle = {Automata, {Languages} and {Programming}},
+	booktitle = {Automata, Languages and Programming},
 	publisher = {Springer},
 	author = {Dwork, Cynthia},
[…]
