+++ README.md
@@ -43,4 +43,3 @@ Note that the `matplotlib` Python package must be available through `PyCall`.
 
 The package is developed and has been tested on Julia 1.6 through 1.10.
 
-[![codecov.io](http://codecov.io/github/goualard-f/MicroFloatingPoints.jl/coverage.svg?branch=main)](http://codecov.io/github/goualard-f/MicroFloatingPoints.jl?branch=main)
