### 2024-07-26 #7030 @matbesancon
although one thing to note: I'm not a numerical analysis person so I won't be able to judge the package or contribution in that regard
