### 2024-07-29 #7030 @danielskatz
@mkitti - I was looking for a third reviewer because of the comment from @matbesancon
> although one thing to note: I'm not a numerical analysis person so I won't be able to judge the package or contribution in that regard

and you were suggested to me as having a good background in Julia and in CS in general
