### 2024-07-29 #7050 @editorialbot
[REVIEW]: MicroFloatingPoints.jl: providing very small IEEE 754-compliant floating-point types
**Submitting author:** @goualard-f (<a href="http://orcid.org/0000-0002-1798-1568">Frederic Goualard</a>)
**Repository:** https://github.com/goualard-f/MicroFloatingPoints.jl.git
**Branch with paper.md** (empty if default branch): main
**Version:** joss2024_accepted
