### 2024-07-29 goualard-f/MicroFloatingPoints.jl#6 @matbesancon
Global variable and stateful package
Issue coming from the review of https://github.com/openjournals/joss-reviews/issues/7050#issuecomment-2254864152

One surprise that is not super idiomatic for Julia packages is seeing that the package is stateful, with the `inexact()` function depending on the package state.

# Short term

One of the drawbacks is that it means the package is probably not thread-safe? This could be documented in the readme if not fixed.
The other thing is that the state is kept in a global variable, which have noticeably poor performance in Julia and are avoided in package code, see https://docs.julialang.org/en/v1/manual/performance-tips/#Avoid-untyped-global-variables

# Long term

Think of a way to allow checked and unchecked computations? That would return whether the computation was exact or not for the checked version only, and the "standard" number API otherwise
