### 2024-07-29 goualard-f/MicroFloatingPoints.jl#8 @matbesancon
Removing Documenter from dependencies
Documenter should be a dependency for the docs/Project.toml only, not in the main package, see:
https://github.com/JuliaLang/Example.jl/blob/master/docs/Project.toml

In the context of the review: https://github.com/openjournals/joss-reviews/issues/7050
