### 2024-07-30 #7050 @dannys4
In addition to being able to replicate the results presented in the documentation and use these types accordingly, it seems I can easily and successfully employ these types in functions I've written previously without any adjustments (in accordance with the goals of JuliaLang), so I'd consider the functionality phenomenal. While I might recommend automating the tests on the GitHub CI, all tests (which seem very comprehensive) do pass locally.

The documentation on the repo is very thorough and incredibly well written! A few minor points on the "tour":
-  I think there should be `using PyCall; plt=pyimport("matplotlib.pyplot")` (or `using PyPlot` or whatever it is) included at the top for users who are perhaps unfamiliar with the PyCall terminology.
- It seems you are missing a `plt` for `imshow` in the tour regarding the plot "Exhaustive search for rounded sums in Floatmu{2,2}".
- I believe you're missing a `using Distributions` in the last code box of the tour
- While there's developer docs, it seems like you're missing contributor guidelines (per the checklist).
Again, great job, and I wish all developers paid as much attention to detail as you.

The document is largely well-written. The only (excessively) minor comment is that "fractional" needs to be spell checked on the caption of Figure 1.

After these minor points above are addressed, I'd be more than happy to recommend acceptance.

As a final note, in the spirit of JOSS' suggestion we stay objective, this has no bearing on acceptance or rejection; however, I'd urge you to consider using Julia [package extensions](https://julialang.org/blog/2023/04/julia-1.9-highlights/#package_extensions) for future endeavors (example [here](https://github.com/pebeto/julia_extensions_example)) to decouple "extra dependencies" (particularly plotting capabilities) from the "core competency" of the package. While the documentation makes a great example of how the plotting can be used as an educational and demonstrative tool, this could very well be an optional tool that can be used _on top of_ the package as opposed to an absolutely necessary "strong" dependency.
