### 2024-08-07 #7050 @danielskatz
👋 @goualard-f - Do you have any comments on what @dannys4 brought up? And on the issues that @matbesancon opened?  The JOSS review is an iterative process, where you don't need to wait for reviewers to complete their initial passes, but can respond to them as they make comments/suggestions.
