### 2024-08-08 #7050 @matbesancon
Hi all,
The paper is clear in terms of description of the software. The package is still early maturity as far as Julia packages go, hence the issues I opened on the repository itself. Having performance gotchas seems problematic for fundamental types like floating points.

The paper motivates a use case of the package:

> MicroFloatingPoints.jl was, for example, instrumental in our understanding of the flaws of the algorithms computing
random floating-point numbers by dividing a random integer by a constant (Goualard, 2020):
being able to execute the algorithms for all possible inputs and to quickly display the results
graphically in Jupyter gave us the impetus to demonstrate rigorously that such procedures
cannot ensure an even distribution of the bits in the fractional parts of the random floats,
rendering them useless for applications such as differential privacy

It is not clear to me what the point of the use of Jupyter is here, is it some usability aspects of the library that are highlighted? What changes compared to other floating point types and libraries?
