### 2024-08-13 b9dd981
Fix Project.toml, eliminate unused dependencies, fix docs on package
