+++ Project.toml
@@ -4,13 +4,12 @@ authors = ["Frédéric Goualard <Frederic.Goualard@univ-nantes.fr>"]
 version = "1.5.3"
 
 [deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SuiteSparse_jll = "bea87d4a-7f5b-5778-9afe-8cc45184846c"
 
 [compat]
-Documenter = "^1"
-PyPlot = "^2"
-julia = "^1.10"
+PyPlot = "2"
+Printf = "1.10"
+Random = "1.10"
+julia = "1.10"
