+++ docs/Project.toml
@@ -2,9 +2,10 @@
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+MicroFloatingPoints = "6e0295d3-a2ae-4d9e-ab31-7b55fc0f4333"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 
 [compat]
-julia = "^1.10"
-PyPlot = "^2"
-Documenter = "^1"
+Documenter = "1"
+PyPlot = "2"
+julia = "1.10"
