### 2024-08-13 #7050 @mkitti
The authors provide a package for small IEEE-754 compliant floating point types. This is implemented via parameterized `Floatmu{szE,szf}` type where `szE` are the number of bits representing the exponent and `szf` are the number of bits representing the fractional part. The total number of bits used is `1 + szE + szf`.

`Floatmu` stores this information in two 32-bit integers as mentioned in a footnote in the documentation. This is somewhat perplexing since often the goal of creating such minifloats is to minimize the storage needs of the numbers as to store more values in memory. The first of the two 32-bit integers stores the bits of the floating point value itself. The second stores an inexact flag which can take values -1, 0, or 1. The type of the first value could be parameterized while the second value could be made smaller to use a type such as a `Int8`.

That is rather the current definition of the `Floatmu` type as follows.

```julia
struct Floatmu{szE, szf} <: AbstractFloat
    v::UInt32
    inexact::Int32
end
```

The type could be made more flexible in the following manner:

```julia
struct Floatmu{szE, szf, T} <: AbstractFloat
    v::T
    inexact::Int8
end
```

The proposed `Floatmu{1, 6, UInt8}` would only require two bytes rather than eight bytes in the current implementation.

Additionally a global 32-bit `inexact_flag` is used. This global state creates a thread safety issue should these numerical types be used in a multi-threaded environment. Perhaps one way this could be addressed is by using [task local storage](https://docs.julialang.org/en/v1/base/parallel/#Base.task_local_storage-Tuple{Any}).

The paper and documentation do not mention the Julia package [BFloat16s.jl](https://github.com/JuliaMath/BFloat16s.jl) although the author does reference BFloat. The author should reference this package and explain the relationship of this work to that prior work as part of explaining the State of the Field.

The overall paper is quite short with a single figure showing a flaw discovered by using the tools within the paper. In contrast the documentation is extensive with multiple figures and a guided tour. The paper could be extended to include a more thorough description of the package and perhaps include some elements from the documentation and guided tour.

There are several software engineering issues that the other reviewers have pointed out. I have submitted some pull requests to the package to try to alleviate some of these issues. The issues include

* Unused dependencies
* Confusion in the structure of the package and documentation environments
* Inclusion of Python and matplotlib as strong rather than weak dependencies

I am still working through the numerical analysis and my review is ongoing.

Overall, the package seems mainly situated as a package meant to analyze the properties of small floating point types rather than providing practical minifloats for use in computational algorithms. Practical use of the package could be improved by addressing the software engineering issues, the global `inexact` flag, and the size of the `Floatmu` type. While the work could be evaluated and published purely as an analysis package, I highly recommend that the authors consider some changes to improve the practical use of the package.
