### 2024-08-13 a8b0dba
Corrected alignment problems in headers (issue #7)
