+++ src/MFPRandom.jl
@@ -1,22 +1,22 @@
 # MFPRandom --
 #
-#	Copyright 2019--2023 University of Nantes, France.
+# Copyright 2019--2024 University of Nantes, France.
 #
-#	This file is part of the MicroFloatingPoints library.
+# This file is part of the MicroFloatingPoints library.
 #
-#	The MicroFloatingPoints library is free software; you can redistribute it and/or
-# modify it under the terms of the GNU Lesser General Public License as published
-# by the Free Software Foundation; either version 3 of the License, or (at your
-#	option) any later version.
+# The MicroFloatingPoints library is free software; you can redistribute
+# it and/or modify it under the terms of the GNU Lesser General Public License
+# as published by the Free Software Foundation; either version 3 of the
+# License, or (at your option) any later version.
+#
+# The MicroFloatingPoints library is distributed in the hope that it will be
+# useful, but	WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# See the GNU General Public License for more details.
 #	
-#	The MicroFloatingPoints library is distributed in the hope that it will be useful,
-# but	WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
-#	or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
-#	for more details.
-#	
-#	You should have received copies of the GNU General Public License and the
-#	GNU Lesser General Public License along with the MicroFloatingPoints Library.
-# If not,	see https://www.gnu.org/licenses/.
+# You should have received copies of the GNU General Public License and the
+# GNU Lesser General Public License along with the MicroFloatingPoints Library.
+# If not, see https://www.gnu.org/licenses/.
 
 module MFPRandom
 
