### 2024-08-13 goualard-f/MicroFloatingPoints.jl#8 @goualard-f
Done, thanks to @mkitti PR.
