### 2024-08-14 #7050 @goualard-f
@dannys4

> I think there should be using PyCall; plt=pyimport("matplotlib.pyplot") (or using PyPlot or whatever it is) included at the top for users who are perhaps unfamiliar with the PyCall terminology.

Added some text at the beginning of Section **Graphics with MicroFloatingPoints.MFPPlot** to remind the user to add `using PyPlot` when necessary.

> It seems you are missing a plt for imshow in the tour regarding the plot "Exhaustive search for rounded sums in Floatmu{2,2}".

Corrected, even though it did not seem to hinder proper compilation.

> I believe you're missing a using Distributions in the last code box of the tour

Corrected (it was performed *under the hood*).

> While there's developer docs, it seems like you're missing contributor guidelines (per the checklist).

There is a section called **Contributing and reporting issues** at the beginning of the documentation but it does not say much, I am afraid, apart from requiring pull requests to contribute.

> The document is largely well-written. The only (excessively) minor comment is that "fractional" needs to be spell checked  on the caption of Figure 1.

Corrected.

> As a final note, in the spirit of JOSS' suggestion we stay objective, this has no bearing on acceptance or rejection; however, I'd urge you to consider using Julia [package extensions](https://julialang.org/blog/2023/04/julia-1.9-highlights/#package_extensions) for future endeavors (example [here](https://github.com/pebeto/julia_extensions_example)) to decouple "extra dependencies" (particularly plotting capabilities) from the "core competency" of the package. While the documentation makes a great example of how the plotting can be used as an educational and demonstrative tool, this could very well be an optional tool that can be used on top of the package as opposed to an absolutely necessary "strong" dependency.

I was not aware of the existence of *package extensions*. Thanks for pointed them out. I will be sure to investigate them.

Thank you very much for your detailed remarks. I have committed the modifications to the `joss2024` branch.
