### 2024-08-14 138eea8
Taking into account @matbesancon remarks. Removal of the global variable for the inexact flag.
