### 2024-08-14 goualard-f/MicroFloatingPoints.jl#6 @goualard-f
Corrected by removing the global variable and using functions with closure.
