### 2024-08-14 #7050 @goualard-f
@matbesancon

> The paper is clear in terms of description of the software. The package is still early maturity as far as Julia packages go, hence the issues I opened on the repository itself. Having performance gotchas seems problematic for fundamental types like floating points.

Two of the issues have been closed, thanks to @mkitti contributions.

I have removed the global variable containing the *inexact flag* and replaced it by functions with closures. I have also added some text at the beginning of the documentation to make it clear to the prospective user that the package is not about performances but about the versatility of easily defining new floating-point types with different precisions and ranges, something that is not offered by packages such as [`BFloat16s.jl`](https://github.com/JuliaMath/BFloat16s.jl) and other packages with a fixed format.

> The paper motivates a use case of the package:
>
>    MicroFloatingPoints.jl was, for example, instrumental in our understanding of the flaws of the algorithms computing
>    random floating-point numbers by dividing a random integer by a constant (Goualard, 2020):
>    being able to execute the algorithms for all possible inputs and to quickly display the results
>    graphically in Jupyter gave us the impetus to demonstrate rigorously that such procedures
>    cannot ensure an even distribution of the bits in the fractional parts of the random floats,
>    rendering them useless for applications such as differential privacy

>It is not clear to me what the point of the use of Jupyter is here, is it some usability aspects of the library that are highlighted? What changes compared to other floating point types and libraries?

I have modified the wording in the paper to make it clearer that Jupyter only brings to the table an integrated environment in which it is easier to test many snippets of code and see the results, thanks to the scripting nature of Julia.

Thank you for your remarks that help make this package better!
