### 2024-08-14 #7050 @goualard-f
@mkitti First, thank you very much for your contribution to this package. For the time being, I have merged your PRs into the `joss2024` branch. Everything should be ultimately merged to the `main` branch.

> `Floatmu` stores this information in two 32-bit integers as mentioned in a footnote in the documentation. This is somewhat perplexing since often the goal of creating such minifloats is to minimize the storage needs of the numbers as to store more values in memory. The first of the two 32-bit integers stores the bits of the floating point value itself. The second stores an inexact flag which can take values -1, 0, or 1. The type of the first value could be parameterized while the second value could be made smaller to use a type such as a `Int8`.
>
> That is rather the current definition of the `Floatmu` type as follows.
>
> ```julia
> struct Floatmu{szE, szf} <: AbstractFloat
>     v::UInt32
>     inexact::Int32
> end
> ```
>
> The type could be made more flexible in the following manner:
>
> ```julia
> struct Floatmu{szE, szf, T} <: AbstractFloat
>     v::T
>     inexact::Int8
> end
> ```

As you state later on yourself, the use case for this package is not the same as, e.g., the BFloat16s.jl package you refer to. The MicroFloatingPoints.jl package is often used for exploratory purpose in algorithms, not necessarily machine learning ones, for which it is not clear what is the impact of precision and range. Performances are traded for versatility: to my knowledge, no other package offers the possibility to choose both the precision and the range (e.g., BFloat16s.jl is stuck with the characteristics of Google Brain's 16 bit floats).

I do not understand how you would use the third parameter `T` in the definition of `Floatmu`. My understanding is that it would make the code of the package much harder to write as the algorithms used currently know in advance the size of the integer used to represent a minifloat.

I do not have enough knowledge of the internal representation adopted by Julia to represent a `Floatmu` to assess the impact of using an `Int8` for `inexact` instead of an `Int32`. Wouldn't it be a problem alignment-wise?

> The proposed `Floatmu{1, 6, UInt8}` would only require two bytes rather than eight bytes in the current implementation.
>
> Additionally a global 32-bit `inexact_flag` is used. This global state creates a thread safety issue should these numerical types be used in a multi-threaded environment. Perhaps one way this could be addressed is by using [task local storage](https://docs.julialang.org/en/v1/base/parallel/#Base.task_local_storage-Tuple%7BAny%7D).

The global variable has been removed and replaced by functions with closure.

>
> The paper and documentation do not mention the Julia package [BFloat16s.jl](https://github.com/JuliaMath/BFloat16s.jl) although the author does reference BFloat. The author should reference this package and explain the relationship of this work to that prior work as part of explaining the State of the Field.
>

I have added a reference to the package in the paper and made it clear that both packages do not have the same use-case.

> The overall paper is quite short with a single figure showing a flaw discovered by using the tools within the paper. In contrast the documentation is extensive with multiple figures and a guided tour. The paper could be extended to include a more thorough description of the package and perhaps include some elements from the documentation and guided tour.
>

My understanding was that the size of the paper should be kept at a minimum since the documentation is the primary source of information. Maybe the editor (@danielskatz) can chip in on that point?

> There are several software engineering issues that the other reviewers have pointed out. I have submitted some pull requests to the package to try to alleviate some of these issues. The issues include
>
>     * Unused dependencies

Should be solved by your PR

>     * Confusion in the structure of the package and documentation environments

I am not sure I understand what you are refering to here.

>     * Inclusion of Python and matplotlib as strong rather than weak dependencies

The package is currently supposed to offer graphical functionnalities as first class citizens, which, I believe makes PyPlot a strong dependency. @dannys4 suggested to use *package extensions*, which I plan to investigate in the near future.

>
> I am still working through the numerical analysis and my review is ongoing.
>
> Overall, the package seems mainly situated as a package meant to analyze the properties of small floating point types rather than providing practical minifloats for use in computational algorithms. Practical use of the package could be improved by addressing the software engineering issues, the global `inexact` flag, and the size of the `Floatmu` type. While the work could be evaluated and published purely as an analysis package, I highly recommend that the authors consider some changes to improve the practical use of the package.
