### 2024-08-15 #7050 @mkitti
Thank you for your reponse to my comments. Thus far, I have only posted my initial impression and critiques. I have not yet suggested specific actions. I will break my notes and responses here into several comments.

> As you state later on yourself, the use case for this package is not the same as, e.g., the BFloat16s.jl package you refer to. The MicroFloatingPoints.jl package is often used for exploratory purpose in algorithms, not necessarily machine learning ones, for which it is not clear what is the impact of precision and range. Performances are traded for versatility: to my knowledge, no other package offers the possibility to choose both the precision and the range (e.g., BFloat16s.jl is stuck with the characteristics of Google Brain's 16 bit floats).

I think it would be fine to consider this primarily as an analysis package, but I would appreciate if this was better documented upfront.

Another that comes to mind is https://github.com/JuliaMath/FixedPointNumbers.jl, which has an underlying type parameter as well as a parameter to control how many bits are committed to the fractional portion of a fixed point number.

> I do not understand how you would use the third parameter `T` in the definition of `Floatmu`. My understanding is that it would make the code of the package much harder to write as the algorithms used currently know in advance the size of the integer used to represent a minifloat.

Julia type parameterization system allows one to relatively easily generate generic code. For example, instead of hard coding `UInt32` below...

```julia
# Mask to retrieve the fractional part (internal use)
significand_mask(::Type{Floatmu{szE,szf}}) where {szE, szf}  = UInt32((UInt32(1) << szf) - 1)
# Mask to retrieve the exponent part (internal use)
exponent_mask(::Type{Floatmu{szE,szf}}) where {szE, szf} = UInt32((UInt32(1) << UInt32(szE))-1) << UInt32(szf)
# Mask to retrieve the sign part (internal use)
sign_mask(::Type{Floatmu{szE,szf}}) where {szE, szf} = UInt32(1) << (UInt32(szE)+UInt32(szf))
```

... we could replace UInt32 by the parameter `T`.

```julia
# Mask to retrieve the fractional part (internal use)
significand_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T}  = T((T(1) << szf) - 1)
# Mask to retrieve the exponent part (internal use)
exponent_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T} = T((T(1) << T(szE))-1) << T(szf)
# Mask to retrieve the sign part (internal use)
sign_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T} = T(1) << (T(szE)+T(szf))
```

If you wanted to how many bits a concrete had for storage, you could use a function like `storage_bits` below.

```
julia> storage_bits(::Type{<: Floatmu{<: Any, <: Any,T}}) where T = sizeof(T)*8
storage_bits (generic function with 1 method)

julia> storage_bits(Floatmu{5,10,UInt32})
32
```

I would also like to point out that the functions above all compile down to methods that just return constant values.

```
julia> @code_llvm significand_mask(Floatmu{5,10,UInt32})
;  @ REPL[7]:1 within `significand_mask`
define i32 @julia_significand_mask_180() #0 {
top:
  ret i32 1023
}
```

If you have a function that maybe strongly depends on the storage type being `UInt32` you could just implement the method for that parameter. A transitional approach might be to do the following.

```julia
struct FloatmuGeneric{szE, szf, T} <: AbstractFloat
    v::T
    inexact::Int8
    # constructors ...
end

const Floatmu{szE, szf} = FloatmuGeneric{szE, szf, UInt32}
```

Now all of your existing code is properly scoped to the storage type being `UInt32`, but you could generalizing the code to arbitrary unsigned integer type `T` where it would be easy to do so.

> I do not have enough knowledge of the internal representation adopted by Julia to represent a `Floatmu` to assess the impact of using an `Int8` for `inexact` instead of an `Int32`. Wouldn't it be a problem alignment-wise?

Julia structs follow the C convention. You are correct that keeping `v` as a `UInt32` and `inexact` as a `Int8` would not save any space due to padding. However, the size could be smaller for smaller storage types.

```julia-repl
julia> sizeof(Floatmu{8,14,UInt32})
8

julia> sizeof(Floatmu{8,14,UInt16})
4

julia> sizeof(Floatmu{8,14,UInt8})
2
```

I do not see any code where alignment would be a major issue. Please correct me if I am wrong.
