### 2024-08-19 66381b4
Added text to documentation and paper to make it clear that the package is ill-suited to applications with massive amounts of data.
