### 2024-08-19 #7050 @goualard-f
@mkitti

> I think it would be fine to consider this primarily as an analysis package, but I would appreciate if this was better documented upfront.

I have modified the introduction of the online documentation to make it clearer and to stress the fact that all formats are represented as a pair of 32 bit integers (see also below). I have also added some text to that effect at the end of the paper (all modifications pushed to the `joss2024` branch).

> Another that comes to mind is https://github.com/JuliaMath/FixedPointNumbers.jl, which has an underlying type parameter as well as a parameter to control how many bits are committed to the fractional portion of a fixed point number.

I did not know this package. However, it believe it is less relevant than, say, the `BFloat16s.jl` package you mentioned earlier as fixed-point numbers have an entirely different error profile (the error being of a fixed size throughout the range).

> > I do not understand how you would use the third parameter `T` in the definition of `Floatmu`. My understanding is that it would make the code of the package much harder to write as the algorithms used currently know in advance the size of the integer used to represent a minifloat.
>
> Julia type parameterization system allows one to relatively easily generate generic code. For example, instead of hard coding `UInt32` below...
>
> ```julia
> # Mask to retrieve the fractional part (internal use)
> significand_mask(::Type{Floatmu{szE,szf}}) where {szE, szf}  = UInt32((UInt32(1) << szf) - 1)
> # Mask to retrieve the exponent part (internal use)
> exponent_mask(::Type{Floatmu{szE,szf}}) where {szE, szf} = UInt32((UInt32(1) << UInt32(szE))-1) << UInt32(szf)
> # Mask to retrieve the sign part (internal use)
> sign_mask(::Type{Floatmu{szE,szf}}) where {szE, szf} = UInt32(1) << (UInt32(szE)+UInt32(szf))
> ```
>
> ... we could replace UInt32 by the parameter `T`.
>
> ```julia
> # Mask to retrieve the fractional part (internal use)
> significand_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T}  = T((T(1) << szf) - 1)
> # Mask to retrieve the exponent part (internal use)
> exponent_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T} = T((T(1) << T(szE))-1) << T(szf)
> # Mask to retrieve the sign part (internal use)
> sign_mask(::Type{Floatmu{szE,szf,T}}) where {szE, szf,T} = T(1) << (T(szE)+T(szf))
> ```
>
> If you wanted to how many bits a concrete had for storage, you could use a function like `storage_bits` below.
>
> ```
> julia> storage_bits(::Type{<: Floatmu{<: Any, <: Any,T}}) where T = sizeof(T)*8
> storage_bits (generic function with 1 method)
>
> julia> storage_bits(Floatmu{5,10,UInt32})
> 32
> ```
>
> I would also like to point out that the functions above all compile down to methods that just return constant values.
>
> ```
> julia> @code_llvm significand_mask(Floatmu{5,10,UInt32})
> ;  @ REPL[7]:1 within `significand_mask`
> define i32 @julia_significand_mask_180() #0 {
> top:
>   ret i32 1023
> }
> ```
>
> If you have a function that maybe strongly depends on the storage type being `UInt32` you could just implement the method for that parameter. A transitional approach might be to do the following.
>
> ```julia
> struct FloatmuGeneric{szE, szf, T} <: AbstractFloat
>     v::T
>     inexact::Int8
>     # constructors ...
> end
>
> const Floatmu{szE, szf} = FloatmuGeneric{szE, szf, UInt32}
> ```
>
> Now all of your existing code is properly scoped to the storage type being `UInt32`, but you could generalizing the code to arbitrary unsigned integer type `T` where it would be easy to do so.
>

That is an interesting generalization as it could potentially allow to compute with floating-point formats with a bit more precision and range than `Float32`, which is presently the limit of the package. With `Float64` as the support for computation ---as it is now--- we could have floats with a 26 bit precision instead of only 23 at most. In the other direction, we could divide by 2 or 4 the size of the representation of the types when they are very small. From my (albeit limited) interaction with users, I am not yet convinced that this is a requested feature, however since the package is primarily used for preliminary investigations with a limited amount of data.
