### 2024-08-20 #7050 @danielskatz
Thanks to everyone for this nice discussion so far
