### 2024-08-27 5f27561
Implement package extensions for PyPlot and PythonPlot as weak deps
