### 2024-08-27 #7050 @mkitti
I have created https://github.com/goualard-f/MicroFloatingPoints.jl/pull/11 that turns PyPlot into a weak dependency via the [package extension](https://pkgdocs.julialang.org/v1/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)) mechanism.

This also allows the more modern [PythonPlot.jl](https://github.com/JuliaPy/PythonPlot.jl) as a weak dependency to be used which depends on [PythonCall.jl](https://github.com/JuliaPy/PythonCall.jl) and [micromamba](https://mamba.readthedocs.io/en/latest/user_guide/micromamba.html) on a per-project basis.

Overall, the dependencies of the core package are reduced those `Random` and `Printf` which are both in the standard library. Via the extension mechansim, additional plotting backends could be implemented. This could include [Plots.jl](https://docs.juliaplots.org/stable/) or [Makie.jl](https://github.com/MakieOrg/Makie.jl) based plots.
