### 2024-08-27 e7afcfc
Remove PyPlot.jl from make.jl for documentation
