### 2024-08-27 5cbcf67
Force Julia to install Python via miniforge
