### 2024-08-27 d1c3b04
Try some block-level setup for doc tests
