+++ src/MFPUtils.jl
@@ -34,14 +34,13 @@ For this function, the rightmost bit of the binary representation of a `Floatmu`
 
 # Examples
 
-```jldoctest; setup = :(using MicroFloatingPoints, MicroFloatingPoints.MFPUtils)
+```jldoctest
 julia> join(string.(reverse(vertical_popcount(Floatmu{2,2}[1.5])))) == bitstring(Floatmu{2,2}(1.5))
 true
 ```
 Note that, in the preceding example, we have to revert the array obtained from `vertical_popcount` because the number of times bit `i` is `1` is saved at position `i`. As a consequence, the value for the rightmost bit of a `Floatmu` appears at the leftmost position of the counting array.
 
-```jldoctest; setup = :(using MicroFloatingPoints, MicroFloatingPoints.MFPUtils)
-
+```jldoctest
 julia> println(vertical_popcount(Floatmu{2,2}[0.25,1.5,3.0]))
 [1, 2, 1, 1, 0]
 ```
