### 2024-08-31 #7050 @danielskatz
@goualard-f - I also notice that the only item @dannys4 hasn't checked of is related to community guidelines. Can you comment on how your software repository handles this? Or improve it?
