### 2024-09-02 546406e
Modified documentation to take into accound modularity regarding PyPlot/PythonPlot.
Co-authored-by: Mark Kittisopikul <mkitti@users.noreply.github.com>
