+++ docs/src/index.md
@@ -61,7 +61,11 @@ Or, equivalently, via the Pkg API:
 julia> import Pkg; Pkg.add("MicroFloatingPoints")
 ```
 
-Note that the `matplotlib` Python package must be available through `PyCall`.
+Note that the `matplotlib` Python package must be available through `PyCall` or `PythonCall`.
+This is usually setup when installing [PyPlot.jl](https://github.com/JuliaPy/PyPlot.jl)
+or [PythonPlot.jl](https://github.com/JuliaPy/PythonPlot.jl), which are optional (weak) dependencies
+that provide plotting backends via package extensions.
+
 
 
 ## Tests
