### 2024-09-02 259d5b9
Added contribution guidelines
