+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+# Contributing
+
+## Reporting an issue
+
+
+
+## Contributing code
