### 2024-09-02 731fcf7
Added contribution guidelines
