### 2024-09-02 #7050 @goualard-f
@danielskatz
> 👋 @goualard-f - Any thoughts on @mkitti's suggestions above?

I have taken some time to review @mkitti's contribution. Once again, he did a great job that enhances significatively the modularity of the package. I have merged his PR into the `joss2024` branch.
