### 2024-09-04 #7050 @danielskatz
👋 @goualard-f and @dannys4 - based on the comment above, it seems like that branch could be merged into the main branch now? Then the item would be complete for JOSS.
