### 2024-09-04 #7050 @mkitti
@danielskatz , could you respond to the following?

> My understanding was that the size of the paper should be kept at a minimum since the documentation is the primary source of information. Maybe the editor (@danielskatz) can chip in on that point?

My contention is that while the documentation does a good job of introducing the package, the paper does not. In particular, these two webpages and their figures provide a great impression of what this package does.
* https://goualard-f.github.io/MicroFloatingPoints.jl/dev/index.html
* https://goualard-f.github.io/MicroFloatingPoints.jl/dev/guided-tour.html

My recent experience with JOSS suggests that the paper could carry more content such that the paper itself carries a full introduction and a "guided tour" of the capabilities of the package:
* https://joss.theoj.org/papers/10.21105/joss.06441
* https://joss.theoj.org/papers/10.21105/joss.04300
* https://joss.theoj.org/papers/10.21105/joss.03965

In particular, the above papers which I reviewed recently include:
1. Explanations of their mathematical basis if relevant.
2. Example of code of how to use the package.
3. Figures explaining the concepts and generated by the above code.

My opinion is that while the the package and its documentation will serve as a living ongoing project, the paper itself should be a complete description of the package at this point in time. As the author already has the materials available, I think is quite relevant and plausible to ask that these elements also be included in the paper.

For potential users of this package, the paper may be the only description of the software that they have before they try to evaluate if the next project may be more relevant to their needs than this one. Thus, I think the paper should be complete and not reliant on the documentation to give a full impression of the package in totality.
