### 2024-09-04 #7050 @mkitti
The one technical point I am still working through is the numerical analysis portion. I will finish that this week.
