### 2024-09-05 0627add
Added an example of use of the package in the paper. Clarified some limitations due to the implementation.
