+++ JOSS2024/paper.bib
@@ -164,3 +164,19 @@
 	year = {2023},
 	note = {Publisher: Frontiers},
 }
+
+
+@article{martin-dorelIssuesRelatedDouble2013,
+	title = {Some issues related to double rounding},
+	volume = {53},
+	issn = {0006-3835, 1572-9125},
+	url = {http://link.springer.com/10.1007/s10543-013-0436-2},
+	doi = {10.1007/s10543-013-0436-2},
+	language = {en},
+	number = {4},
+	journal = {BIT Numerical Mathematics},
+	author = {Martin-Dorel, Érik and Melquiond, Guillaume and Muller, Jean-Michel},
+	month = dec,
+	year = {2013},
+	pages = {897--924},
+}
