+++ JOSS2024/paper.md
@@ -12,21 +12,69 @@ authors:
 affiliations:
  - name: Nantes Université, École Centrale Nantes, CNRS, LS2N, UMR 6004, Nantes, France
    index: 1
-date: 19 August 2024
+date: 5 September 2024
 bibliography: paper.bib
 ---
 
 # Summary
 
-The IEEE 754 standard defines the representation and the properties of the floating-point numbers used as surrogates for reals in most applications. Usually, programming languages only support the two formats ---corresponding to different ranges and precisions--- implemented in most hardware: `Float32` on 32 bits, and `Float64` on 64 bits (aka, respectively, `float` and `double` in programming languages of  the C family). Machine learning, Computer Graphics, and numerical algorithms analysis all have a need for smaller formats ([minifloats](https://en.wikipedia.org/wiki/Minifloat)), which are often neither supported in hardware, nor are they available as established types in programming languages. The [`MicroFloatingPoints.jl`](https://github.com/goualard-f/MicroFloatingPoints.jl) Julia library offers a parametric type that can be instantiated to compute with IEEE 754-compliant floating-point numbers with varying ranges and precisions (up to and including `Float32`). It also provides the programmer with various means to visualize what is computed.
+The IEEE 754 standard defines the representation and the properties of the floating-point numbers used as surrogates for reals in computer programs. Most programming languages only support the 32-bit (`Float32`) and 64-bit (`Float64`) formats implemented in hardware. Machine learning, Computer Graphics, and numerical algorithms analysis all have a need for smaller formats, which are often neither supported in hardware, nor are they available as established types in programming languages. The [`MicroFloatingPoints.jl`](https://github.com/goualard-f/MicroFloatingPoints.jl) Julia library offers a parametric type that can be instantiated to compute with IEEE 754-compliant floating-point numbers with varying ranges and precisions (up to and including `Float32`). It also provides the programmer with various means to visualize what is computed.
 
 # Statement of need
 
[…]
