### 2024-09-05 #7050 @goualard-f
@danielskatz @matbesancon @mkitti @dannys4
I have modified the article to include a quick tour; I have also added a section to make clearer the limitations of the package.
