### 2024-09-05 #7050 @goualard-f
@matbesancon That's odd since the article was correctly generated when I pushed it (https://github.com/goualard-f/MicroFloatingPoints.jl/actions/runs/10717448528). I had a look at the [log](https://github.com/openjournals/joss-papers/actions/runs/10717549024/job/29717465232) and the error is quite obscure to me:

```
Error running filter /usr/local/share/openjournals/data/filters/prepare-affiliations.lua:
...share/openjournals/data/filters/prepare-affiliations.lua:11: attempt to index a nil value (field 'integer index')
```

Has anybody any experience about that?
