### 2024-09-05 #7050 @goualard-f
@matbesancon I have downloaded locally the whole [script package](https://github.com/openjournals/inara) to compile the paper on my computer. I am able to replicate the error only when I request an output in the `crossref` format. PDF generates fine.

What is more curious is that I am able to replicate the exact same error when I try to create a `crossref` version of the [example](https://joss.readthedocs.io/en/latest/example_paper.html) paper. I would be inclined to believe there has been some change in the expected format that is not yet documented.
@danielskatz Do you know of anybody who could confirm this?
