### 2024-09-05 #7050 @goualard-f
@matbesancon @danielskatz I have been able to replicate the same error with an [article ](https://joss.theoj.org/papers/10.21105/joss.06964) that has been published by JOSS on 27 August 2024. That strengthen my belief that the code of my article might not be at fault.
