### 2024-09-05 #7050 @danielskatz
Sorry - see https://github.com/openjournals/joss/issues/1353 - hopefully this will be fixed shortly
