### 2024-09-05 #7050 @goualard-f
@danielskatz @matbesancon That is a relief. In the meantime, the modified paper can be read from [here](https://github.com/goualard-f/MicroFloatingPoints.jl/actions/runs/10717448528/artifacts/1895307480).
