### 2024-09-05 #7050 @dannys4
Pending the merge of the `joss2024` branch, this all looks good and I think that Mark's paper feedback greatly improved the result. The only small piece of feedback is that it seems the "Statement of Need" section trails off at the end. Was this intended?
