### 2024-09-05 #7050 @goualard-f
@dannys4
> Pending the merge of the `joss2024` branch, this all looks good and I think that Mark's paper feedback greatly improved the result. The only small piece of feedback is that it seems the "Statement of Need" section trails off at the end. Was this intended?

Yes it was. The last sentence is an open enumeration of equivalences between `Floatmu{}` instantiations with existing types. I could modify the sentence if you believe it is not adequate.
