### 2024-09-06 bd5b8c0
Add multitasking tests for inexact
