### 2024-09-06 d7b7a53
Fix documentation CI failure
