### 2024-09-06 #7050 @mkitti
In response to @matbesancon 's point about type instability and thread safety of the inexact flag, https://github.com/goualard-f/MicroFloatingPoints.jl/issues/6 , a revision was made moving the flag into a `let` block creating several functional closures.

However, both type instability and thread safety issues remain with the current implementation. The macro `@code_warntype` from the standard library InteractiveUtils.jl package provides insight into Julia's type inferrence.  Below we see that Julia is still unable to infer the return type of `MicroFloatingPoints.inexact()`.

```julia-repl
julia> @code_warntype MicroFloatingPoints.inexact()
MethodInstance for MicroFloatingPoints.inexact()
  from inexact() @ MicroFloatingPoints ~/.julia/dev/MicroFloatingPoints/src/Floatmu.jl:72
Arguments
  #self#::Core.Const(MicroFloatingPoints.inexact)
Locals
  inexact_flag::Union{}
Body::Any
1 ─ %1 = $(QuoteNode(Core.Box(false)))
│   %2 = Core.isdefined(%1, :contents)::Core.Const(true)
└──      goto #3 if not %2
2 ─      goto #4
3 ─      Core.Const(Core.NewvarNode(:(inexact_flag)))
└──      Core.Const(:(inexact_flag))
4 ┄ %7 = Core.getfield(%1, :contents)::Any
└──      return %7
```

The immediate fix to this involves adding a type assertion as follows.
```diff
diff --git a/src/Floatmu.jl b/src/Floatmu.jl
index 43bf3cb..2b61461 100644
--- a/src/Floatmu.jl
+++ b/src/Floatmu.jl
@@ -44,7 +44,7 @@ export reinterpret
 export fractional_even

-let inexact_flag = false
+let inexact_flag::Bool = false
     global set_inexact_or, inexact, reset_inexact
     function set_inexact_or(v)
         inexact_flag = inexact_flag || v
```

The thread safety issues remain because there remains no barrier to multiple threads attempting to access `inexact` at the same time, creating a race condition. One solution to this might be to declare `inexact` as an [atomic variable](https://docs.julialang.org/en/v1/manual/multi-threading/#Atomic-Operations). However, it is not clear to me that this would actually allow for multithread concurrency.

Consider the following example.

```julia-repl
julia> begin
           MicroFloatingPoints.reset_inexact()
           task_A = Threads.@spawn begin
               sleep(.2)
               MicroFloatingPoints.set_inexact_or(true)
               sleep(.3)
               @info "Task A, expecting true" MicroFloatingPoints.inexact()
           end
           task_B = Threads.@spawn begin
               for i in 0.1:0.1:0.5
                   @info "Task B, $i sec " MicroFloatingPoints.inexact()
                   if MicroFloatingPoints.inexact()
                       MicroFloatingPoints.reset_inexact()
                   end
                   sleep(.1)
               end
           end
           wait(task_A)
       end
┌ Info: Task B, 0.1 sec
└   MicroFloatingPoints.inexact() = false
┌ Info: Task B, 0.2 sec
└   MicroFloatingPoints.inexact() = false
┌ Info: Task B, 0.3 sec
└   MicroFloatingPoints.inexact() = true
┌ Info: Task B, 0.4 sec
└   MicroFloatingPoints.inexact() = false
┌ Info: Task B, 0.5 sec
└   MicroFloatingPoints.inexact() = false
┌ Info: Task A, expecting true
└   MicroFloatingPoints.inexact() = false
```

Above `inexact_flag` is set to `true` on Task A. However, Task B sees the `inexact_flag` set to `true`, despite no inexact operations happening on that Task, and then resets it. By the end of `Task A`, the `inexact_flag` is no longer set.

I have added a version of the above demonstration in commit [bd5b8c0](https://github.com/goualard-f/MicroFloatingPoints.jl/pull/12/commits/bd5b8c0440ee65db606aca434dfd2a603bf6f491) as part of https://github.com/goualard-f/MicroFloatingPoints.jl/pull/12 . These tests currently [fail](https://github.com/goualard-f/MicroFloatingPoints.jl/actions/runs/10740260205/job/29787917001?pr=12).

Here I recommend the use of `task_local_storage`. This allows each `Task` to have their own `inexact_flag` meaning that multiple parallel operations can occur in parallel while independent tracking inexactness. A major change here is that the `inexact_flag` must be checked before a `Task` completes. I have implemented this in https://github.com/goualard-f/MicroFloatingPoints.jl/pull/12/commits/e92ad3a4576db8ac61387e5a21b9de599d0713a7 as part of https://github.com/goualard-f/MicroFloatingPoints.jl/pull/12.
