### 2024-09-06 e61d8cc
Modified documentation to make it clear that the inexact flag is now local to each task.
