### 2024-09-06 #7050 @goualard-f
Thank you very much @mkitti . Once again, you went above and beyond. I have merged your PR and modified the documentation to make it clear that the inexact flag is now local to each task.
