### 2024-09-09 #7050 @matbesancon
In the paper, in order to encourage and show best practices, it might be better to use a const for aliases:
```
const MuFP = Floatmu{E,f} # The new format
```

I'm also not sure what is meant with the comment "The new format"?
The vector T could probably be a vector of Int instead of floats

Is `plt` defined by PyPlot or is it an alias for `PyPlot`?
