### 2024-09-09 b7e8c20
Modified example in paper.
