### 2024-09-09 #7050 @goualard-f
> In the paper, in order to encourage and show best practices, it might be better to use a const for aliases:
>
> ```
> const MuFP = Floatmu{E,f} # The new format
> ```

Right. I guess that should also be the case for `E`, `f` and `p`. I have modified the article accordingly.

> I'm also not sure what is meant with the comment "The new format"? The vector T could probably be a vector of Int instead of floats

`T` is now an array of `UInt32`. I have also changed the comment to '`New IEEE 754 type`'.

> Is `plt` defined by PyPlot or is it an alias for `PyPlot`?

As per the [documentation](https://github.com/JuliaPy/PyPlot.jl?tab=readme-ov-file#exported-functions) for `PyPlot`, it seems to be defined as an alias for `matplotlib.pyplot`.
