### 2024-09-17 5f6bba3
Corrected typos in the bibliography. Polished some sentences in the text. Added an acknowledgment section at the end.
