+++ JOSS2024/paper.bib
@@ -22,7 +22,7 @@
 	doi = {10.1109/ARITH.2013.22},
 	booktitle = {Proceedings of the 2013 {IEEE} 21st Symposium on Computer Arithmetic},
 	publisher = {IEEE Computer Society},
-	author = {Lefevre, Vincent},
+	author = {Lefèvre, Vincent},
 	month = apr,
 	year = {2013},
 	pages = {99--106},
@@ -30,7 +30,7 @@
 
 
 @article{rumpIEEE754PrecisionBase2016,
-	title = {{IEEE754} Precision-\$k\$ base-\${\textbackslash}beta\$ Arithmetic Inherited by Precision-\$m\$ Base-\${\textbackslash}beta\$ Arithmetic for \$k {\textless} m\$},
+	title = {{IEEE754} Precision-$k$ base-$\beta$ Arithmetic Inherited by Precision-$m$ Base-$\beta$ Arithmetic for $k<m$},
 	volume = {43},
 	issn = {0098-3500},
 	url = {http://doi.org/10.1145/2785965},
@@ -53,7 +53,7 @@
 	language = {en},
 	number = {4},
 	journal = {ACM Transactions on Mathematical Software},
-	author = {Flegar, Goran and Scheidegger, Florian and Novaković, Vedran and Mariani, Giovani and Tom´s, Andrés E. and Malossi, A. Cristiano I. and Quintana-Ortí, Enrique S.},
+	author = {Flegar, Goran and Scheidegger, Florian and Novaković, Vedran and Mariani, Giovani and Tomás, Andrés E. and Malossi, A. Cristiano I. and Quintana-Ortí, Enrique S.},
 	month = dec,
 	year = {2019},
 	pages = {1--23},
