### 2024-09-17 #7050 @danielskatz
@goualard-f - it looks like all reviewers are satisfied with this. Next, I will proofread it, and then tell you what else is needed in the JOSS process.
