### 2024-09-17 b505f40
minor changes in paper.md
