### 2024-09-17 #7050 @danielskatz
@goualard-f - I've suggested some minor changes in https://github.com/goualard-f/MicroFloatingPoints.jl/pull/14 - please merge this, or let me know what you disagree with, then we can continue to publication.
