### 2024-09-18 #7050 @goualard-f
@danielskatz
>     * Make a tagged release of your software, and list the version tag of the archived version here.

[joss2024_accepted](https://github.com/goualard-f/MicroFloatingPoints.jl/releases/tag/joss2024_accepted)

>     * Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

https://zenodo.org/records/13777404

>     * Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

I have added @mkitti to the contributors since he has provided a lot of code during the review process. I hope this does not contravene your statement above.

>     * Please list the DOI of the archived version here.

 10.5281/zenodo.13777404

> I can then move forward with accepting the submission.

I would like to thank @matbesancon , @dannys4 , @mkitti , and yourself for your dedication to make the library and the paper better throughout the review process. I am honestly impressed by the level of commitment of all involved.
