+++ v2.0.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"DocTestSetup = quote\n    using MicroFloatingPoints\n\tusing MicroFloatingPoints.MFPPlot, MicroFloatingPoints.MFPRandom, PyPlot\nend\nCurrentModule = MicroFloatingPoints","category":"page"},{"location":"guided-tour.html#A-guided-tour","page":"A Guided Tour","title":"A guided tour","text":"","category":"section"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"The MicroFloatingPoints package is organized into four modules:","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"MicroFloatingPoints: the main module containing the definition of the parameterized type Floatmu and the associated methods;\nMicroFloatingPoints.MFPUtils: a module providing miscellaneous utility functions for the Floatmu type;\nMicroFloatingPoints.MFPPlot: a module offering various graphical ways to display Floatmu floating-point numbers;\nMicroFloatingPoints.MFPRandom: the module overloading Random.rand to produce Floatmu random values.","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"After having correctly installed the package (see Installation), we start our tour by loading the MicroFloatingPoints module:","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"using MicroFloatingPoints","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"We can now define a new floating-point type MuFP with 2 bits for the exponent (the first parameter)  and 2 bits for the fractional part (the second parameter):","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"MuFP = Floatmu{2,2}","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"Such a type is very limited, and a call to floatmax will give us the largest finite float representable:","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"floatmax(MuFP)","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"Conversely, we can obtain the smallest positive float in the MuFP format with the μ method:","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"μ(MuFP)","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"Note that this value is a subnormal number, which is different and smaller than the smallest normal float, obtained by calling floatmin:","category":"page"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"floatmin(MuFP)","category":"page"},{"location":"guided-tour.html#Graphics-with-MicroFloatingPoints.MFPPlot","page":"A Guided Tour","title":"Graphics with MicroFloatingPoints.MFPPlot","text":"","category":"section"},{"location":"guided-tour.html","page":"A Guided Tour","title":"A Guided Tour","text":"The MicroFloatingPoints package offers several graphical functionalities that are all available in MicroFloatingPoints.MFPPlot when the plotting package PyPlot is also loaded:","category":"page"},{"location":"guided-tour.html
[…]
