### 2024-09-18 #7050 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
