### 2024-09-18 0f0bda1
Added citation refs.
