### 2024-09-18 #7050 @danielskatz
👋 @goualard-f - Congratulations on your publication!!

And thanks to @matbesancon, @dannys4, and @mkitti for reviewing!
JOSS depends on volunteers and couldn't succeed without you
