### 2024-09-18 #7050 @danielskatz
@goualard-f - I now see that you encoded part of your title, which has made https://joss.theoj.org/papers/10.21105/joss.07050 and the associated metadata not very nice. I'm sorry I missed this while proofreading.

Can you please remove the encoding in the title in the .md file, and then I will reprocess the paper.
