### 2024-09-18 #7050 @danielskatz
No - JOSS doesn't need the final paper source to be in the repo, as what is archived for the paper is the PDF and the JATS versions, along with the generated metadata.
