### 2024-09-18 #7050 @goualard-f
@danielskatz I have pushed the corrected `paper.md` to `main`. Is that sufficient?
