### 2024-09-18 #7050 @danielskatz
Thanks @goualard-f - now https://doi.org/10.21105/joss.07050 looks good, and I'll close this issue now
